Moved new functionality into it's own function.
