+++ packages/core/src/operator/index.ts
@@ -10,3 +10,4 @@ export * from "./claimRewardBulk.js";
 export * from "./submitBulkAssertions.js";
 export * from "./claimBulkSubmissionRewards.js";
 export * from "./operatorRuntime.js";
+export * from "./checkRefereeBulkSubmissionCompatible.js";
